+++ designer/client/src/components/modals/DeployProcessDialog.tsx
@@ -5,8 +5,9 @@ import React, { useCallback, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
 
-import type { NodesDeploymentData } from "../../http/HttpService";
-import { getProcessName, getProcessVersionId } from "../../reducers/selectors/graph";
+import type { NodesDeploymentData, ScenarioSource} from "../../http/HttpService";
+import { ScenarioSourceType } from "../../http/HttpService";
+import { getGraph, getProcessName, getProcessVersionId, isSaveDisabled } from "../../reducers/selectors/graph";
 import { getFeatureSettings } from "../../reducers/selectors/settings";
 import type { WindowKind } from "../../windowManager";
 import { PromptContent } from "../../windowManager";
@@ -23,6 +24,7 @@ export type ToggleProcessActionModalData = {
         processVersionId: ProcessVersionId,
         comment: string,
         nodeData?: NodesDeploymentData,
+        scenarioSource?: ScenarioSource,
     ) => Promise<ScenarioActionResult>;
     displayWarnings?: boolean;
     actionName?: string;
@@ -39,9 +41,15 @@ export function DeployProcessDialog(props: WindowContentProps<WindowKind, Toggle
     const [validationError, setValidationError] = useState("");
     const featureSettings = useSelector(getFeatureSettings);
     const deploymentCommentSettings = featureSettings.deploymentCommentSettings;
+    const unsavedScenarioSource: ScenarioSource = {
+        type: ScenarioSourceType.FROM_GRAPH,
+        scenarioGraph: useSelector(getGraph)?.scenario?.scenarioGraph,
+    };
+    const savedScenarioSource: ScenarioSource = { type: ScenarioSourceType.LATEST_VERSION };
+    const scenarioSource: ScenarioSource = useSelector(isSaveDisabled) ? savedScenarioSource : unsavedScenarioSource;
 
     const confirmAction = useCallback(async () => {
-        const response = await action(processName, processVersionId, comment);
+        const response = await action(processName, processVersionId, comment, null, scenarioSource);
         switch (response.scenarioActionResultType) {
             case ScenarioActionResultType.Success:
             case ScenarioActionResultType.UnhandledError:
@@ -54,7 +62,7 @@ export function DeployProcessDialog(props: WindowContentProps<WindowKind, Toggle
                 console.log("Unexpected result type:", response.scenarioActionResultType);
                 break;
         }
-    }, [action, comment, processName, props, processVersionId]);
+    }, [action, comment, processName, props, processVersionId, scenarioSource]);
 
     const { t } = useTranslation();
     const buttons: WindowButtonProps[] = useMemo(