+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/process/deployment/DeploymentServiceSpec.scala
@@ -2,6 +2,7 @@ package pl.touk.nussknacker.ui.process.deployment
 
 import cats.implicits.toTraverseOps
 import cats.instances.list._
+import io.circe.{parser, Decoder}
 import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach, OptionValues}
 import org.scalatest.LoneElement._
 import org.scalatest.matchers.should.Matchers
@@ -17,9 +18,17 @@ import pl.touk.nussknacker.engine.api.deployment.ProcessStateDefinitionManager.S
 import pl.touk.nussknacker.engine.api.deployment.ScenarioActionName.{Cancel, Deploy}
 import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
 import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus.ProblemStateStatus
+import pl.touk.nussknacker.engine.api.graph.ScenarioGraph
 import pl.touk.nussknacker.engine.api.process._
 import pl.touk.nussknacker.engine.build.ScenarioBuilder
-import pl.touk.nussknacker.engine.deployment.{DeploymentId, ExternalDeploymentId}
+import pl.touk.nussknacker.engine.deployment.{
+  DeploymentId,
+  ExternalDeploymentId,
+  FromGraph,
+  LatestVersion,
+  ScenarioSource
+}
+import pl.touk.nussknacker.engine.spel.SpelExtension.SpelExpresion
 import pl.touk.nussknacker.test.{EitherValuesDetailedMessage, NuScalaTestAssertions, PatientScalaFutures}
 import pl.touk.nussknacker.test.base.db.WithHsqlDbTesting
 import pl.touk.nussknacker.test.base.it.WithClock
@@ -42,7 +51,7 @@ import pl.touk.nussknacker.ui.process.ScenarioQuery
 import pl.touk.nussknacker.ui.process.deployment.scenariostatus.FragmentStateException
 import pl.touk.nussknacker.ui.process.periodic.flink.FlinkClientStub
 import pl.touk.nussknacker.ui.process.repository.{CommentValidationError, DBIOActionRunner}
-import pl.touk.nussknacker.ui.process.repository.ProcessRepository.CreateProcessAction
+import pl.touk.nussknacker.ui.process.repository.ProcessRepository.{CreateProcessAction, UpdateProcessAction}
 import pl.touk.nussknacker.ui.security.api.LoggedUser
 
 import java.time.Instant
@@ -130,7 +139,8 @@ class DeploymentServiceSpec
           RunDeploymentCommand(
             CommonCommandData(scenario, None, user),
             StateRestoringStrategy.RestoreStateFromReplacedJobSavepoint,
-            NodesDeploymentData.empty
+            NodesDeploymentData.empty,
+            scenarioSource = LatestVersion,
           )
         )
         .failed
@@ -154,7 +164,8 @@ class DeploymentServiceSpec
       RunDeploymentCommand(
         CommonCommandData(scenario, None, user),
         StateRestoringStrategy.RestoreStateFromReplacedJobSavepoint,
-        NodesDeploymentData.empty
+        NodesDeploymentData.empty,
+        scenarioSource = LatestVersion,
       )
     )
 
@@ -184,7 +195,8 @@ class DeploymentServiceSpec
             RunDeploymentCommand(
               CommonCommandData(scenario, Comment.from("samplePattern"), user),
               StateRestoringStrategy.RestoreStateFromReplacedJobSavepoint,
-              NodesDeploymentData.empty
+              NodesDeploymentData.empty,
+              scenarioSource = LatestVersion,
             )
           )
           .futureValue
@@ -232,7 +244,8 @@ class DeploymentServiceSpec
             RunDeploymentCommand(
               CommonCommandData(scenario, None, user),
               StateRestoringStrategy.RestoreStateFromReplacedJobSavepoint,
-              NodesDeploymentData.empty
+              NodesDeploymentData.empty,
+              scenarioSource = LatestVersion,
             )
           )
           .futureValue
@@ -334,7 +347,8 @@ class DeploymentServiceSpec
             RunDeploymentCommand(
               CommonCommandData(scenario, None, user),
               StateRestoringStrategy.RestoreStateFromReplacedJobSavepoint,
-              NodesDeploymentData.empty
+              NodesDeploymentData.empty,
+              scenarioSource = LatestVersion,
             )
           )
           .futureValue
@@ -376,7 +390,8 @@ class DeploymentServiceSpec
             RunDeploymentCommand(
               CommonCommandData(scenario, None, user),
               StateRestoringStrategy.RestoreStateFromReplacedJobSavepoint,
-              NodesDeploymentData.empty
+              NodesDeploymentData.empty,
+              scenarioSource = LatestVersion,
             )
           )
           .failed
@@ -769,7 +784,8 @@ class DeploymentServiceSpec
             RunDeploymentCommand(
               CommonCommandData(scenario, None, user),
               StateRestoringStrategy.RestoreStateFromReplacedJobSavepoint,
-              NodesDeploymentData.empty
+              NodesDeploymentData.empty,
+              scenarioSource = LatestVersion,
             )
           )
           .futureValue
@@ -1211,10 +1227,112 @@ class DeploymentServiceSpec
     }
   }
 
+  "should deploy scenario taking scenario from different sources" when {
+    "should deploy scenario latest version" in {
+      val scenarioName = generateScenarioName()
+      val result =
+        deploymentManager1.withScenarioStateStatus(ProcessName(scenarioName), SimpleStateStatus.NotDeployed) {
+          deployExampleScenario(scenarioName, LatestVersion)
+        }
+
+      result should not be None
+    }
+
+    "should deploy scenario from given scenario graph" in {
+      val scenarioName = generateScenarioName()
+      val scenario     = prepareScenario(scenarioName)
+      val scenarioGraphJson =
+        """
+          |{
+          |  "properties" : {
+          |    "additionalFields" : {
+          |      "description" : null,
+          |      "properties" : {
+          |        "parallelism" : "",
+          |        "spillStateToDisk" : "true",
+          |        "useAsyncInterpretation" : "",
+          |        "checkpointIntervalInSeconds" : ""
+          |      },
+          |      "metaDataType" : "StreamMetaData",
+          |      "showDescription" : false
+          |    }
+          |  },
+          |  "nodes" : [
+          |    {
+          |      "id" : "source",
+          |      "ref" : {
+          |        "typ" : "barSource",
+          |        "parameters" : [
+          |        ]
+          |      },
+          |      "additionalFields" : null,
+          |      "type" : "Source"
+          |    },
+          |    {
+          |      "id" : "newVariableAdded",
+          |      "varName" : "newVar",
+          |      "value" : {
+          |        "language" : "spelTemplate",
+          |        "expression" : "updated process with new variable"
+          |      },
+          |      "additionalFields" : null,
+          |      "type" : "Variable"
+          |    },
+          |    {
+          |      "id" : "sink",
+          |      "ref" : {
+          |        "typ" : "barSink",
+          |        "parameters" : [
+          |        ]
+          |      },
+          |      "endResult" : null,
+          |      "isDisabled" : null,
+          |      "additionalFields" : null,
+          |      "type" : "Sink"
+          |    }
+          |  ],
+          |  "edges" : [
+          |    {
+          |      "from" : "source",
+          |      "to" : "newVariableAdded",
+          |      "edgeType" : null
+          |    },
+          |    {
+          |      "from" : "newVariableAdded",
+          |      "to" : "sink",
+          |      "edgeType" : null
+          |    }
+          |  ],
+          |  "stickyNotes" : [
+          |  ]
+          |}
+          |""".stripMargin
+      val scenarioGraph = parser.parse(scenarioGraphJson).flatMap(Decoder[ScenarioGraph].decodeJson).rightValue
+
+      val result = deploymentManager1.withScenarioStateStatus(scenario.name, SimpleStateStatus.NotDeployed) {
+        deployExistingScenario(scenario, FromGraph(scenarioGraph))
+      }
+      val lastFinishedAction = actionRepository.getFinishedProcessActions(scenario.id, None).dbioActionValues.head
+
+      result should not be None
+      lastFinishedAction.actionName shouldBe ScenarioActionName("DEPLOY")
+      lastFinishedAction.processVersionId shouldBe VersionId(2)
+      lastFinishedAction.state shouldBe ProcessActionState.Finished
+    }
+  }
+
   private def deployExampleScenario(
-      scenarioName: String = generateScenarioName()
+      scenarioName: String = generateScenarioName(),
+      scenarioSource: ScenarioSource = LatestVersion,
   ): Option[ExternalDeploymentId] = {
     val scenario = prepareScenario(scenarioName)
+    deployExistingScenario(scenario, scenarioSource)
+  }
+
+  private def deployExistingScenario(
+      scenario: ProcessIdWithName,
+      scenarioSource: ScenarioSource = LatestVersion,
+  ): Option[ExternalDeploymentId] = {
     deploymentManager1
       .withWaitForDeployFinish(scenario.name, result = Some(ExternalDeploymentId("1"))) {
         deploymentService
@@ -1222,7 +1340,8 @@ class DeploymentServiceSpec
             RunDeploymentCommand(
               CommonCommandData(scenario, None, user),
               StateRestoringStrategy.RestoreStateFromReplacedJobSavepoint,
-              NodesDeploymentData.empty
+              NodesDeploymentData.empty,
+              scenarioSource = scenarioSource,
             )
           )
           .futureValue
@@ -1238,7 +1357,8 @@ class DeploymentServiceSpec
             RunRedeploymentCommand(
               CommonCommandData(scenarioId, None, user),
               StateRestoringStrategy.RestoreStateFromReplacedJobSavepoint,
-              NodesDeploymentData.empty
+              NodesDeploymentData.empty,
+              scenarioSource = LatestVersion,
             )
           )
           .futureValue
@@ -1380,6 +1500,25 @@ class DeploymentServiceSpec
       .saveNewProcess(action)
       .map(_.value.processId)
       .map(ProcessIdWithName(_, ProcessName(scenarioName)))
+      .dbioActionValues
+  }
+
+  private def updateScenario(scenario: ProcessIdWithName): Option[VersionId] = {
+    val canonicalProcess = ScenarioBuilder
+      .streaming(scenario.name.value)
+      .source("source", ProcessTestData.existingSourceFactory)
+      .buildSimpleVariable("newVariableAdded", "newVar", "updated process with new variable".spelTemplate)
+      .emptySink("sink", ProcessTestData.existingSinkFactory)
+    val action = UpdateProcessAction(
+      processId = scenario.id,
+      canonicalProcess = canonicalProcess,
+      comment = Comment.from("Update Comment"),
+      labels = Nil,
+      increaseVersionWhenJsonNotChanged = true
+    )
+    writeProcessRepository
+      .updateProcess(action)
+      .map(_.newVersion)
       .dbioActionValues
   }