+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/process/periodic/flink/PeriodicDeploymentManagerTest.scala
@@ -16,7 +16,7 @@ import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
 import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus.ProblemStateStatus
 import pl.touk.nussknacker.engine.api.process.{ProcessId, ProcessIdWithName, ProcessName, VersionId}
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
-import pl.touk.nussknacker.engine.deployment.{DeploymentData, User}
+import pl.touk.nussknacker.engine.deployment.{DeploymentData, LatestVersion, User}
 import pl.touk.nussknacker.test.PatientScalaFutures
 import pl.touk.nussknacker.test.base.db.WithHsqlDbTesting
 import pl.touk.nussknacker.test.config.WithSimplifiedDesignerConfig.TestProcessingType.Streaming
@@ -140,7 +140,8 @@ class PeriodicDeploymentManagerTest
         RunDeploymentCommand(
           CommonCommandData(ProcessIdWithName(id, processName), None, user),
           RestoreStateFromReplacedJobSavepoint,
-          NodesDeploymentData.empty
+          NodesDeploymentData.empty,
+          scenarioSource = LatestVersion,
         )
       )
       .futureValue