+++ extensions-api/src/main/scala/pl/touk/nussknacker/engine/deployment/ScenarioSource.scala
@@ -1,0 +1,10 @@
+package pl.touk.nussknacker.engine.deployment
+
+import io.circe.generic.JsonCodec
+import io.circe.generic.extras.ConfiguredJsonCodec
+import pl.touk.nussknacker.engine.api.CirceUtil._
+import pl.touk.nussknacker.engine.api.graph.ScenarioGraph
+
+@ConfiguredJsonCodec sealed trait ScenarioSource
+case object LatestVersion                                           extends ScenarioSource
+@JsonCodec final case class FromGraph(scenarioGraph: ScenarioGraph) extends ScenarioSource